server changes

+++ database/migrations/2022_04_27_093801_create_hotels_table.php
@@ -19,14 +19,13 @@ return new class extends Migration
             $table->string('description');
             $table->string('phone');
             $table->float('rate');
-            $table->float('latitude');
-            $table->float('longtude');
+            $table->double('latitude');
+            $table->double('longtude');
             $table->string('address');
             $table->string('services');
             $table->string('hotel_url')->nullable();
             $table->timestamps();
             $table->softDeletes();
-            
         });
     }
